backend/requirements.txt
# --- Framework ---
Django==5.1.4

# --- Base de datos ---
psycopg2-binary==2.9.10     # conector Python ↔ PostgreSQL

# --- Variables de entorno ---
python-decouple==3.8

# --- Admin ---
django-jazzmin==3.0.1

# --- Frontend ---
django-htmx==1.21.0

# --- Exportación ---
openpyxl==3.1.5             # Excel
reportlab==4.2.5            # PDF (facturas)

# --- Servidor de producción ---
gunicorn==21.2.0